corregimos errores
correccion de errores, sacamos set de usuario de perfil, y actualizamos base de datos
