+++ Grupo-14-OO2-2021/src/main/java/com/unla/Grupo14OO22021/converters/PerfilConverter.java
@@ -1,7 +1,5 @@
 package com.unla.Grupo14OO22021.converters;
 
-
-
 import org.springframework.stereotype.Component;
 
 import com.unla.Grupo14OO22021.entities.Perfil;
