+++ Grupo-14-OO2-2021/src/main/java/com/unla/Grupo14OO22021/converters/UsuarioConverter.java
@@ -1,5 +1,7 @@
 package com.unla.Grupo14OO22021.converters;
 
+
+
 import org.springframework.stereotype.Component;
 
 
@@ -9,12 +11,16 @@ import com.unla.Grupo14OO22021.models.UsuarioModel;
 @Component("usuarioConverter")
 public class UsuarioConverter {
 	public UsuarioModel entityToModel(Usuario usuario) {
-		return new UsuarioModel(usuario.getIdUsuario(),usuario.getNombre(),usuario.getApellido(),usuario.getTipoDocumento(),usuario.getNumeroDocumento(),
+		return new UsuarioModel(usuario.getIdUsuario(),usuario.getNombre(),
+				usuario.getApellido(),usuario.getTipoDocumento(),usuario.getNumeroDocumento(),
 				usuario.getEmail(),usuario.getNomUsuario(),usuario.getContraseña());
 	}
 	
 	public Usuario modelToEntity(UsuarioModel usuarioModel) {
-		return new Usuario(usuarioModel.getIdUsuario(),usuarioModel.getNombre(),usuarioModel.getApellido(),usuarioModel.getTipoDocumento(),
-				usuarioModel.getNumeroDocumento(),usuarioModel.getEmail(),usuarioModel.getNomUsuario(),usuarioModel.getContraseña());
+		PerfilConverter perfilConverter = new PerfilConverter();
+		return new Usuario(usuarioModel.getIdUsuario(), usuarioModel.getNombre(), 
+				usuarioModel.getApellido(), usuarioModel.getTipoDocumento(), 
+				usuarioModel.getNumeroDocumento(), usuarioModel.getEmail(), 
+				usuarioModel.getNomUsuario(), usuarioModel.getContraseña(), perfilConverter.modelToEntity(usuarioModel.getPerfil()));
 	}
 }
