+++ Grupo-14-OO2-2021/src/main/java/com/unla/Grupo14OO22021/entities/Usuario.java
@@ -35,4 +35,137 @@ public class Usuario {
 	@ManyToOne
 	@JoinColumn (name="idPerfil")
     private Perfil perfil;
+
+	
+	
+	public Usuario() {
+		super();
+	}
+
+
+
+	public Usuario(int idUsuario, String nombre, String apellido, String tipoDocumento, int numeroDocumento,
+			String email, String nomUsuario, String contraseña, Perfil perfil) {
+		super();
+		this.idUsuario = idUsuario;
+		this.nombre = nombre;
+		this.apellido = apellido;
+		this.tipoDocumento = tipoDocumento;
+		this.numeroDocumento = numeroDocumento;
+		this.email = email;
+		this.nomUsuario = nomUsuario;
+		this.contraseña = contraseña;
+		this.perfil = perfil;
+	}
+
+
+
+	public int getIdUsuario() {
+		return idUsuario;
+	}
+
+
+
+	public void setIdUsuario(int idUsuario) {
+		this.idUsuario = idUsuario;
+	}
+
+
+
+	public String getNombre() {
+		return nombre;
+	}
+
+
+
+	public void setNombre(String nombre) {
+		this.nombre = nombre;
+	}
+
+
+
+	public String getApellido() {
+		return apellido;
+	}
+
+
+
+	public void setApellido(String apellido) {
+		this.apellido = apellido;
+	}
+
+
+
+	public String getTipoDocumento() {
+		return tipoDocumento;
+	}
+
+
+
+	public void setTipoDocumento(String tipoDocumento) {
+		this.tipoDocumento = tipoDocumento;
+	}
+
+
+
+	public int getNumeroDocumento() {
+		return numeroDocumento;
+	}
+
+
+
+	public void setNumeroDocumento(int numeroDocumento) {
+		this.numeroDocumento = numeroDocumento;
+	}
+
+
+
+	public String getEmail() {
+		return email;
+	}
+
+
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+
+
+	public String getNomUsuario() {
+		return nomUsuario;
+	}
+
+
+
+	public void setNomUsuario(String nomUsuario) {
+		this.nomUsuario = nomUsuario;
+	}
+
+
+
+	public String getContraseña() {
+		return contraseña;
+	}
+
+
+
+	public void setContraseña(String contraseña) {
+		this.contraseña = contraseña;
+	}
+
+
+
+	public Perfil getPerfil() {
+		return perfil;
+	}
+
+
+
+	public void setPerfil(Perfil perfil) {
+		this.perfil = perfil;
+	}
+	
+	
+	
 }
